+++ src/main/java/de/arrobait/antlers/editor/actions/backspace/AntlersDoubleQuotedStringBackspaceHandler.java
@@ -1,0 +1,7 @@
+package de.arrobait.antlers.editor.actions.backspace;
+
+public class AntlersDoubleQuotedStringBackspaceHandler extends AntlersStringBackspaceHandler{
+    public AntlersDoubleQuotedStringBackspaceHandler() {
+        super('\"');
+    }
+}