+++ src/main/java/de/arrobait/antlers/editor/actions/backspace/AntlersSingleQuotedStringBackspaceHandler.java
@@ -1,0 +1,7 @@
+package de.arrobait.antlers.editor.actions.backspace;
+
+public class AntlersSingleQuotedStringBackspaceHandler extends AntlersStringBackspaceHandler {
+    public AntlersSingleQuotedStringBackspaceHandler() {
+        super('\'');
+    }
+}