+++ src/main/java/de/arrobait/antlers/highlighter/AntlersTemplateHighlighter.java
@@ -1,13 +1,17 @@
 package de.arrobait.antlers.highlighter;
 
-import com.intellij.lang.html.HTMLLanguage;
+import com.intellij.lang.Language;
 import com.intellij.openapi.editor.colors.EditorColorsScheme;
 import com.intellij.openapi.editor.ex.util.LayerDescriptor;
 import com.intellij.openapi.editor.ex.util.LayeredLexerEditorHighlighter;
+import com.intellij.openapi.fileTypes.FileType;
+import com.intellij.openapi.fileTypes.FileTypes;
 import com.intellij.openapi.fileTypes.SyntaxHighlighter;
 import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.psi.templateLanguages.TemplateDataLanguageMappings;
+import de.arrobait.antlers.AntlersLanguage;
 import de.arrobait.antlers.psi.AntlersTypes;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -17,8 +21,20 @@ public class AntlersTemplateHighlighter extends LayeredLexerEditorHighlighter {
         // create the main highlighter
         super(new AntlersSyntaxHighlighter(), colors);
 
-        SyntaxHighlighter htmlHighlighter = SyntaxHighlighterFactory.getSyntaxHighlighter(HTMLLanguage.INSTANCE, project, virtualFile);
-        LayerDescriptor htmlLayer = new LayerDescriptor(htmlHighlighter, "");
-        registerLayer(AntlersTypes.OUTER_CONTENT, htmlLayer);
+        FileType type = null;
+        if (project == null || virtualFile == null) {
+            type = FileTypes.PLAIN_TEXT;
+        } else {
+            Language language = TemplateDataLanguageMappings.getInstance(project).getMapping(virtualFile);
+            if (language != null) {
+                type = language.getAssociatedFileType();
+            }
+            if (type == null) {
+                type = AntlersLanguage.getDefaultTemplateLang();
+            }
+        }
+
+        SyntaxHighlighter outerHighlighter = SyntaxHighlighterFactory.getSyntaxHighlighter(type, project, virtualFile);
+        registerLayer(AntlersTypes.OUTER_CONTENT, new LayerDescriptor(outerHighlighter, ""));
     }
 }